backend-46/core/requirements.txt
asgiref==3.8.1
barely-json==1.0.0
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
dnspython==2.7.0
drf-yasg==1.21.10
flake8==7.3.0
inflection==0.5.1
iniconfig==2.1.0
Markdown==3.8
mccabe==0.7.0
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.1
pymongo==4.13.2
pyparsing==3.2.3
pytest==8.4.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.31.0
sqlparse==0.5.3
uritemplate==4.2.0
djangorestframework-simplejwt==5.5.0